refactor(core): compose logical steps over scoped attempts
